noise linting

+++ server/c_lib/map_gen/interpolator.c
@@ -143,7 +143,7 @@ void _perlin_interp1(float final[], int x, int x_interval,
     // generate anchor points
     int i;
     for (i=0; i<nx; i++) {
-        points[i] = perlin1(((i+1)/fnx)*xnoise_scale, rep, base);   
+        points[i] = perlin1(((float)(i+1)/fnx)*xnoise_scale, rep, base);   
     }
 
     _interp1(final, points, x, x_interval);
@@ -163,7 +163,7 @@ void _rmf_perlin_interp1(float final[], int x, int x_interval,
     // generate anchor points
     int i;
     for (i=0; i<nx; i++) {
-        points[i] = rmf_perlin1(((i+1)/fnx)*xnoise_scale, rep, base);   
+        points[i] = rmf_perlin1(((float)(i+1)/fnx)*xnoise_scale, rep, base);   
     }
 
     _interp1(final, points, x, x_interval);
@@ -278,7 +278,7 @@ void _perlin_interp2(float final[], int x, int y,
     int i,j;
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
-            points[i + nx*j] = perlin2(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, rep_x, rep_y, base);   
+            points[i + nx*j] = perlin2(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, rep_x, rep_y, base);   
         }
     }
 
@@ -303,7 +303,7 @@ void _rmf_perlin_interp2(float final[], int x, int y,
     int i,j;
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
-            points[i + nx*j] = rmf_perlin2(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, rep_x, rep_y, base);   
+            points[i + nx*j] = rmf_perlin2(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, rep_x, rep_y, base);   
         }
     }
     _interp2(final, points, x, y, x_interval, y_interval);
@@ -325,7 +325,7 @@ void _simplex_interp2(float final[], int x, int y,
     int i,j;
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
-            points[i + nx*j] = simplex2(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale);   
+            points[i + nx*j] = simplex2(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale);   
         }
     }
     _interp2(final, points, x, y, x_interval, y_interval);
@@ -347,7 +347,7 @@ void _rmf_simplex_interp2(float final[], int x, int y,
     int i,j;
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
-            points[i + nx*j] = rmf_simplex2(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale);   
+            points[i + nx*j] = rmf_simplex2(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale);   
         }
     }
     _interp2(final, points, x, y, x_interval, y_interval);
@@ -496,7 +496,7 @@ void _perlin_interp3(float final[], int x, int y, int z,
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
             for (k=0; k<nz; k++) {
-                points[i + nx*j + nx*ny*k] = perlin3(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, ((k+1)/fnz)*znoise_scale, rep_x, rep_y, rep_z, base);   
+                points[i + nx*j + nx*ny*k] = perlin3(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, ((float)(k+1)/fnz)*znoise_scale, rep_x, rep_y, rep_z, base);   
             }
         }
     }
@@ -524,7 +524,7 @@ void _rmf_perlin_interp3(float final[], int x, int y, int z,
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
             for (k=0; k<nz; k++) {
-                points[i + nx*j + nx*ny*k] = rmf_perlin3(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, ((k+1)/fnz)*znoise_scale, rep_x, rep_y, rep_z, base);
+                points[i + nx*j + nx*ny*k] = rmf_perlin3(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, ((float)(k+1)/fnz)*znoise_scale, rep_x, rep_y, rep_z, base);
             }
         }
     }
@@ -550,7 +550,7 @@ void _simplex_interp3(float final[], int x, int y, int z,
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
             for (k=0; k<nz; k++) {
-                points[i + nx*j + nx*ny*k] = simplex3(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, ((k+1)/fnz)*znoise_scale);   
+                points[i + nx*j + nx*ny*k] = simplex3(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, ((float)(k+1)/fnz)*znoise_scale);   
             }
         }
     }
@@ -576,7 +576,7 @@ void _rmf_simplex_interp3(float final[], int x, int y, int z,
     for (i=0; i<nx; i++) {
         for (j=0; j<ny; j++) {
             for (k=0; k<nz; k++) {
-                points[i + nx*j + nx*ny*k] = rmf_simplex3(((i+1)/fnx)*xnoise_scale, ((j+1)/fny)*ynoise_scale, ((k+1)/fnz)*znoise_scale);   
+                points[i + nx*j + nx*ny*k] = rmf_simplex3(((float)(i+1)/fnx)*xnoise_scale, ((float)(j+1)/fny)*ynoise_scale, ((float)(k+1)/fnz)*znoise_scale);   
             }
         }
     }
